Save for Estimation not properly working

+++ src/main/java/com/costsestimationbackend/costsestimationbackend/model/Estimation/Estimation.java
@@ -25,7 +25,7 @@ public class Estimation {
     private int estimation;
 
 
-    public long getIdEstimation() {
+    public int getIdEstimation() {
         return idEstimation;
     }
